The total number of characters are available on the route: http://localhost:3001/rickandmorty/character/count
